requirements.txt
ultralytics
supervision==0.24.0
PyYAML
opencv-python-headless
matplotlib
Pillow